ai-engine/requirements.txt
fastapi
uvicorn
python-multipart
openai
google-generativeai
langchain
pydantic
python-dotenv
